pyproject.toml
[tool.poetry]
name = "reseller-api"
version = "0.1.0"
description = "API to manage reseller's cashback"
authors = ["Carlos Gonçalves <carlos.empre@yahoo.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
python-decouple = "^3.5"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.2"
django-environ = "^0.8.1"
drf-yasg = {version = "^1.20.0", extras = ["Werkzeug"]}
djangorestframework-simplejwt = "^5.0.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^4.5.1"
django-extensions = "^3.1.5"
model-bakery = "^1.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
